requirements.txt
# Arquivo com as dependências do projeto.
requests>=2.25.1,<3.0.0
pytest>=6.2.4,<8.0.0
selenium>=4.0.0,<5.0.0
playwright>=1.15.0,<2.0.0
# Dependências do Robot Framework
robotframework>=4.0,<7.0.0
robotframework-requests>=0.9.0,<1.0.0
robotframework-seleniumlibrary>=5.0,<7.0.0
